refactor(container): mv DocUploader to tool scope
