+++ src/containers/editor/PostEditor/EditorFooter.js
@@ -3,7 +3,7 @@ import { pluck } from 'ramda'
 
 import { ICON_CMD } from '@/config'
 
-import DocUploader from '@/containers/DocUploader'
+import DocUploader from '@/containers/tool/DocUploader'
 import Labeler from '@/containers/Labeler'
 import Maybe from '@/components/Maybe'
